New: workhours datepicker inputs replaced with new datepicker
