+++ client/src/components/DatePicker/types.ts
@@ -1,11 +1,11 @@
 export interface IDatePicker {
   disabled?: boolean;
   error?: string;
-  format?: 'MM/dd/yyyy' | 'MM/dd/yyyy  hh:mma';
+  format?: 'MM/dd/yyyy' | 'MM/dd/yyyy  hh:mma' | 'hh:mma';
   fullWidth?: boolean;
+  handleChange: (date: Date) => void;
   icon?: string;
   label: string;
-  onChange: (date: Date) => void;
   name: string;
   value: Date;
 }
