--- client/src/components/DatePickerV2/datePicker.module.css
@@ -1,40 +1,0 @@
-.field {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  gap: 0.6rem;
-
-  height: 9rem;
-  min-width: 20rem;
-}
-
-.iconInputWrapper {
-  position: relative;
-
-  width: 100%;
-}
-
-.datePicker.fullWidth {
-  width: 100%;
-}
-
-.datePicker > div:first-child {
-  padding-left: 4rem;
-}
-
-.iconInputWrapper i {
-  font-size: 1.6rem;
-  margin: 0;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  height: 100%;
-  width: 4rem;
-
-  position: absolute;
-  top: 0;
-  left: 0;
-}
