+++ client/src/components/Input/index.tsx
@@ -1,4 +1,3 @@
-export * from "./Date";
 export * from "./Input";
 export * from "./Profile";
 export * from "./Time";
