+++ client/src/pages/sandbox/index.tsx
@@ -1,4 +1,4 @@
-import { DatePickerV2 as DatePicker, Input } from 'components';
+import { DatePicker, Input } from 'components';
 
 import styles from "./sandbox.module.css";
 import { useState } from 'react';
@@ -8,7 +8,7 @@ export const Sandbox = (): JSX.Element => {
 
   return (
     <div className={ styles.container }>
-      <DatePicker fullWidth icon="calendar outline" label="Date" name="date" onChange={onValueChange} value={value} />
+      <DatePicker fullWidth icon="calendar outline" label="Date" name="date" handleChange={onValueChange} value={value} />
       <Input fullWidth handleChange={() => console.log('')} icon="calendar outline" label="Input" name="input" value="Input" />
     </div>
   );
